Fix up format_data()
There was an issue with format_data() where it was reading the whole
output of redis_read_product_info(). This fix adds logic to just add
the currency information from redis and to set a default string if no
data is read.

This change also updates the name of a function, adds unit tests to address
logic,and fixes an issue with pep8.